# docs/specs/phase-0/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Culinary Blog API
  version: 1.0.0
  description: >-
    Contract triển khai đã hợp nhất từ SRS v1.0.0 và
    00_SRS_Conformance_and_Resolution.md. Mọi lỗi dùng RFC 7807.
servers:
  - url: http://localhost:5000/api/v1
    description: Development
  - url: https://api.culinaryblog.com/api/v1
    description: Production
...
paths:
...
  /categories:
    get:
      tags: [Categories]
      operationId: listCategories
      security: []
      responses:
        '200':
          description: Danh sách category, sắp xếp name tăng dần
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: {$ref: '#/components/schemas/Category'}
    post:
      tags: [Categories]
      operationId: createCategory
      x-required-role: Admin
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CategoryWrite'}
      responses:
        '201': {$ref: '#/components/responses/CategoryResource'}
        '400': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
  /categories/{slug}:
    get:
      tags: [Categories]
      operationId: getCategoryBySlug
      security: []
      parameters:
        - {$ref: '#/components/parameters/Slug'}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
      responses:
        '200':
          description: Category và Published recipes
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CategoryDetailEnvelope'}
        '400': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
  /categories/{id}:
    put:
      tags: [Categories]
      operationId: updateCategory
      x-required-role: Admin
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CategoryWrite'}
      responses:
        '200': {$ref: '#/components/responses/CategoryResource'}
        '400': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
    delete:
      tags: [Categories]
      operationId: deleteCategory
      x-required-role: Admin
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses:
        '204': {description: Category đã soft-delete}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
  /admin/recipes:
    get:
      tags: [Admin]
      operationId: listAllRecipesForAdmin
      x-required-role: Admin
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - {name: status, in: query, schema: {$ref: '#/components/schemas/RecipeStatus'}}
        - {name: authorId, in: query, schema: {type: string}}
      responses:
        '200': {$ref: '#/components/responses/RecipePage'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
  /admin/recipes/{id}:
    get:
      tags: [Admin]
      operationId: getAnyRecipeForAdmin
      x-required-role: Admin
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses:
        '200': {$ref: '#/components/responses/RecipeResource'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
  /recipes:
    get:
      tags: [Recipes]
      operationId: listPublishedRecipes
      security: []
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - {$ref: '#/components/parameters/Sort'}
        - {name: categoryId, in: query, schema: {type: string, format: uuid}}
        - {name: difficulty, in: query, schema: {$ref: '#/components/schemas/RecipeDifficulty'}}
        - {name: maxCookTime, in: query, schema: {type: integer, minimum: 0}}
        - {name: minServings, in: query, schema: {type: integer, minimum: 1}}
        - {name: minPrepTime, in: query, schema: {type: integer, minimum: 1}}
        - {name: maxPrepTime, in: query, schema: {type: integer, minimum: 1}}
      responses:
        '200': {$ref: '#/components/responses/RecipePage'}
        '400': {$ref: '#/components/responses/Problem'}
    post:
      tags: [Recipes]
      operationId: createRecipe
      x-required-role: AuthorOrAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateRecipeRequest'}
      responses:
        '201': {$ref: '#/components/responses/RecipeResource'}
        '400': {$ref: '#/components/responses/Problem'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
...
  /recipes/{slug}:
    get:
      tags: [Recipes]
      operationId: getPublishedRecipeBySlug
      security: []
      parameters: [{$ref: '#/components/parameters/Slug'}]
      responses:
        '200': {$ref: '#/components/responses/RecipeResource'}
        '404': {$ref: '#/components/responses/Problem'}
  /recipes/{id}:
    put:
      tags: [Recipes]
      operationId: updateRecipe
      x-resource-owner-or-admin: true
      parameters:
        - {$ref: '#/components/parameters/Id'}
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateRecipeRequest'}
      responses:
        '200': {$ref: '#/components/responses/RecipeResource'}
        '400': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
    delete:
      tags: [Recipes]
      operationId: deleteRecipe
      x-resource-owner-or-admin: true
      parameters:
        - {$ref: '#/components/parameters/Id'}
        - {$ref: '#/components/parameters/IfMatch'}
      responses:
        '204': {description: Recipe đã soft-delete}
        '400': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
...
  /me/recipes:
    get:
      tags: [MyRecipes]
      operationId: listMyRecipes
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - {name: status, in: query, schema: {$ref: '#/components/schemas/RecipeStatus'}}
      responses:
        '200': {$ref: '#/components/responses/RecipePage'}
        '401': {$ref: '#/components/responses/Problem'}
  /me/recipes/{id}:
    get:
      tags: [MyRecipes]
      operationId: getMyRecipe
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses:
        '200': {$ref: '#/components/responses/RecipeResource'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
  /recipes/{id}/images:
    post:
      tags: [Images]
      operationId: uploadRecipeImage
      x-resource-owner-or-admin: true
      parameters:
        - {$ref: '#/components/parameters/Id'}
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: {type: string, format: binary}
                altText: {type: [string, 'null'], maxLength: 200}
                isPrimary: {type: boolean, default: false}
      responses:
        '201': {$ref: '#/components/responses/ChildMutation'}
        '400': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
        '503': {$ref: '#/components/responses/Problem'}
  /recipes/{id}/images/{imageId}:
    patch:
      tags: [Images]
      operationId: updateRecipeImage
      x-resource-owner-or-admin: true
      parameters:
        - {$ref: '#/components/parameters/Id'}
        - {name: imageId, in: path, required: true, schema: {type: string, format: uuid}}
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                altText: {type: [string, 'null'], maxLength: 200}
                isPrimary: {type: boolean}
                orderIndex: {type: integer, minimum: 0}
      responses:
        '200': {$ref: '#/components/responses/ChildMutation'}
        '400': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
    delete:
      tags: [Images]
      operationId: deleteRecipeImage
      x-resource-owner-or-admin: true
      parameters:
        - {$ref: '#/components/parameters/Id'}
        - {name: imageId, in: path, required: true, schema: {type: string, format: uuid}}
        - {$ref: '#/components/parameters/IfMatch'}
      responses:
        '204': {description: Image metadata deleted; object deletion queued}
        '400': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
  /recipes/{id}/steps:
    post:
      tags: [Steps]
      operationId: createRecipeStep
      x-resource-owner-or-admin: true
      parameters:
        - {$ref: '#/components/parameters/Id'}
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/StepWrite'}
      responses:
        '201': {$ref: '#/components/responses/ChildMutation'}
        '400': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
  /recipes/{id}/steps/{stepId}:
    put:
      tags: [Steps]
      operationId: updateRecipeStep
      x-resource-owner-or-admin: true
      parameters:
        - {$ref: '#/components/parameters/Id'}
        - {name: stepId, in: path, required: true, schema: {type: string, format: uuid}}
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/StepWrite'}
      responses:
        '200': {$ref: '#/components/responses/ChildMutation'}
        '400': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
    delete:
      tags: [Steps]
      operationId: deleteRecipeStep
      x-resource-owner-or-admin: true
      parameters:
        - {$ref: '#/components/parameters/Id'}
        - {name: stepId, in: path, required: true, schema: {type: string, format: uuid}}
        - {$ref: '#/components/parameters/IfMatch'}
      responses:
        '204': {description: Step deleted; remaining steps renumbered}
        '400': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
  /recipes/{id}/ingredients:
    post:
      tags: [Ingredients]
      operationId: createRecipeIngredient
      x-resource-owner-or-admin: true
      parameters:
        - {$ref: '#/components/parameters/Id'}
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/IngredientWrite'}
      responses:
        '201': {$ref: '#/components/responses/ChildMutation'}
        '400': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    Slug: {name: slug, in: path, required: true, schema: {type: string, minLength: 1}}
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    PageSize: {name: pageSize, in: query, schema: {type: integer, minimum: 1, maximum: 50, default: 12}}
...
  responses:
    Problem:
      description: RFC 7807 Problem Details
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ProblemDetails'}
...
    CategoryResource:
      description: Category resource
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties: {data: {$ref: '#/components/schemas/Category'}}
...
  schemas:
...
    Category:
      type: object
      required: [id, name, slug, orderIndex, recipeCount]
      properties:
        id: {type: string, format: uuid}
        name: {type: string, minLength: 2, maxLength: 100}
        slug: {type: string, maxLength: 120}
        description: {type: [string, 'null'], maxLength: 2000}
        imageUrl: {type: [string, 'null'], format: uri}
        orderIndex: {type: integer, minimum: 0}
        recipeCount: {type: integer, minimum: 0}
    CategoryWrite:
      type: object
      required: [name]
      properties:
        name: {type: string, minLength: 2, maxLength: 100}
        description: {type: [string, 'null'], maxLength: 2000}
        imageUrl: {type: [string, 'null'], format: uri, maxLength: 500}
        orderIndex: {type: integer, minimum: 0}
    RecipeStatus: {type: string, enum: [draft, published, archived]}
...
    CategoryDetailEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [category, recipes]
          properties:
            category: {$ref: '#/components/schemas/Category'}
            recipes: {type: array, items: {$ref: '#/components/schemas/RecipeSummary'}}
        meta: {$ref: '#/components/schemas/PageMeta'}
...
security:
  - bearerAuth: []